requirements.txt
# GitBridge Requirements
# Task: P20P5S4 - Finalize requirements.txt with working versions
# Date: 2025-06-19

# =============================================================================
# Core Dependencies
# =============================================================================

# Web Framework
Flask==3.0.0
flask-socketio==5.3.6

# HTTP and API
requests==2.31.0
httpx==0.24.1
httpcore==0.16.3
aiohttp==3.9.1

# AI Service Integrations
openai==1.12.0  # GPT-4o integration (P20P2S1)
# groq==0.4.0  # Grok 3 integration (P20P6 - future)

# Data Validation and Serialization
pydantic==2.5.0  # Schema validation (P20P2S2)
python-dateutil==2.8.2
typing-extensions==4.9.0

# =============================================================================
# Security and Authentication
# =============================================================================

# Cryptography and JWT
cryptography==42.0.2
PyJWT==2.8.0

# Environment and Configuration
python-dotenv==1.0.0

# =============================================================================
# Data Storage and Caching
# =============================================================================

# Redis for caching
redis==5.0.1

# =============================================================================
# Testing Framework
# =============================================================================

# Core Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5

# Property-based Testing
hypothesis==6.98.0

# =============================================================================
# Documentation and Logging
# =============================================================================

# Documentation
markdown==3.5.1
mdx_truly_sane_lists==1.3
flasgger==0.9.7.1

# Logging
python-json-logger==2.0.7

# =============================================================================
# Monitoring and Metrics
# =============================================================================

# Metrics Collection
prometheus-client==0.19.0
statsd==4.0.1

# System Monitoring
psutil==5.9.8

# =============================================================================
# Utilities and Helpers
# =============================================================================

# Time and Date
pytz==2024.1

# Retry Logic
backoff==2.2.1

# Progress Bars
tqdm==4.66.1

# WebSocket Support
websockets==12.0

# =============================================================================
# Development Tools
# =============================================================================

# Code Formatting
black==24.1.1

# Linting
pylint==3.0.3

# Type Checking
mypy==1.8.0

# =============================================================================
# Version Compatibility Notes
# =============================================================================

# Critical Version Locks:
# - openai==1.12.0: Compatible with Python 3.13
# - httpx==0.24.1: Required for OpenAI SDK compatibility
# - httpcore==0.16.3: Required for httpx compatibility
# - pydantic==2.5.0: Schema validation for GPT4oEventSchema

# Python Version: 3.13.3
# Tested and confirmed working on macOS 22.6.0

# =============================================================================
# Future Dependencies (P20P6 - P20P7)
# =============================================================================

# Grok 3 Integration (P20P6)
# groq==0.4.0

# SmartRouter Arbitration (P20P7)
# fastapi==0.104.1
# uvicorn==0.24.0

# Database Integration (Future)
# sqlalchemy==2.0.23
# alembic==1.13.1

# =============================================================================
# Installation Instructions
# =============================================================================

# For development:
# pip install -r requirements.txt

# For production:
# pip install -r requirements.txt --no-dev

# For testing:
# pip install -r requirements.txt[test] 